Fix service definition in the CLI fat JAR.
The JAR was getting multiple entries for the same file which resulted
in only one plugin being visible. Ideally they would be merged, but
I'm not sure how to do that in Gradle, so for now just move the
service provider metadata into the CLI package since we don't need the
build plugins to work separately anyway.