Require config packages in CMake.
If the system has a Find$DEP.cmake available, CMake might prefer that
definition to ours. It almost certainly won't contain libraries for
Android, and it may not even expose the same interface. It may not
necessarily find *any* libraries at all.

Enforce that a config-style package is loaded rather than the
module-style package (see
https://cmake.org/cmake/help/latest/command/find_package.html).