Add note for implicit API level in duplicates.
The pull-up done for the API level prior to the first supported API
level for the ABI is done silently, so users that aren't aware of this
behavior are likely to be confused by the apparent non-duplicates in
their duplicate lists. Add a note to the error message to make it
clear.

When we eventually break the schema we should just explicitly disallow
API levels that were never supported for the given ABI.

Fixes https://github.com/google/prefab/issues/100