Skip modules that have no match.
A package might contain modules that are only compatible with some
configurations. The motivating example of this is where a package
contains a shared library that uses c++_shared, but the user is using
c++_static. In this case, the one incompatible module shouldn't
preclude the user from using other modules in the package.

I don't really like the implementation here. The decision of whether
incompatible modules should be skipped or cause an error should not be
made by each build system plugin, but it'll take some rearranging of
the build system interface API to make that the CLI's responsibility.
I'm making this quick fix for our in-tree build systems for now, and
will the organization when we have more changes that justify a major
release.