Fix "Windows can't clone repo"
Windows fails to clone the repo because of the directory name
"android.". To make it possible for Windows to clone the repo, this
change renames "android." to "android" and then renames the directory
before running the test.

To avoid polluting the resource directory, a copy of the tests data is
made in a temp directory.